+++ libplip/Core/PlipCore.h
@@ -39,7 +39,7 @@ namespace Plip {
         void Step();
 
         void ClearActiveBreakpoint() { m_activeBreakpoint = UINT64_MAX; }
-        virtual void Delta(long ns) = 0;
+        virtual void Delta(double ns) = 0;
         uint64_t GetActiveBreakpoint() const { return m_activeBreakpoint; }
         virtual std::map<std::string, std::map<std::string, DebugValue>> GetDebugInfo() const = 0;
         [[nodiscard]] PlipMemoryMap* GetMemoryMap() const { return m_memory; }
@@ -64,7 +64,7 @@ namespace Plip {
 
         std::set<uint64_t> m_breakpoints {};
 
-        long m_singleStepTime = {};
+        double m_singleStepTime = {};
 
     private:
         uint64_t m_activeBreakpoint = UINT64_MAX;
